Refactored TPSTable#getTPSData to a query
